Addressed feedback from simon
