+++ samples/azure/azure-service-fabric-routing/Core_6/Shared/Shared.csproj
@@ -101,6 +101,7 @@
     <Compile Include="ReceiverSideDistribution\Configuration\PartitionAwareReceiverSideDistributionConfiguration.cs" />
     <Compile Include="ReceiverSideDistribution\PartitionMappingFailedException.cs" />
     <Compile Include="EndpointPartitioning\HardcodeReplyToAddressToLogicalAddress.cs" />
+    <Compile Include="SagaEntry.cs" />
     <Compile Include="SenderSideDistribution\Configuration\PartitionAwareSenderSideDistributionConfiguration.cs" />
     <Compile Include="SenderSideDistribution\Configuration\SenderSideDistributionExtensions.cs" />
     <Compile Include="ReceiverSideDistribution\Pipeline\DistributeMessagesBasedOnHeader.cs" />
