+++ samples/azure/azure-service-fabric-routing/Core_6/ZipCodeVoteCount/EndpointCommunicationListener.cs
@@ -6,7 +6,6 @@ using Microsoft.ServiceFabric.Data;
 using Microsoft.ServiceFabric.Data.Collections;
 using Microsoft.ServiceFabric.Services.Communication.Runtime;
 using NServiceBus;
-using NServiceBus.Persistence.ServiceFabric;
 
 public class EndpointCommunicationListener :
     ICommunicationListener
@@ -72,33 +71,5 @@ public class EndpointCommunicationListener :
     {
         // Fire & Forget Close
         CloseAsync(CancellationToken.None);
-    }
-}
-
-namespace NServiceBus.Persistence.ServiceFabric
-{
-    using System;
-    using System.Runtime.Serialization;
-
-    [DataContract(Namespace = "NServiceBus.Persistence.ServiceFabric", Name = "SagaEntry")]
-    sealed class SagaEntry : IExtensibleDataObject
-    {
-        public SagaEntry(string data, Version sagaTypeVersion, Version persistenceVersion)
-        {
-            Data = data;
-            SagaTypeVersion = sagaTypeVersion;
-            PersistenceVersion = persistenceVersion;
-        }
-
-        [DataMember(Name = "Data", Order = 0)]
-        public string Data { get; private set; }
-
-        [DataMember(Name = "PersistenceVersion", Order = 1)]
-        public Version PersistenceVersion { get; set; }
-
-        [DataMember(Name = "SagaTypeVersion", Order = 2)]
-        public Version SagaTypeVersion { get; set; }
-
-        public ExtensionDataObject ExtensionData { get; set; }
     }
 }
